refactor(tests): extract virt fixtures from tests/conftest.py
Move fixtures from tests/conftest.py into domain-scoped modules
under tests/fixtures/virt/:

- instance_types.py: instance type and VM preference fixtures
- migration.py: migration policy and migrated_vm_multiple_times fixtures
- osinfo.py: libosinfo DB download fixtures
- ssh.py: SSH key generation fixtures
- templates.py: base_templates fixture
- virtual_machines.py: VM creation fixtures, rhel_latest_os_params,
  vm_for_migration_test

Register new modules in pytest_plugins in root conftest.py.
